Finishing core replay playback functionality; adding recording and playback of ability cancellation and hold position; using a more reliable format for ability UIDs; previews are working now

+++ Assets/Scripts/Gameplay/Config/Abilities/Implementation/HealAbilityDataAsset.cs
@@ -33,7 +33,7 @@ namespace Gameplay.Config.Abilities {
 
         public override IAbilityParameters DeserializeParametersFromJson(Dictionary<string, object> json) {
             return new HealAbilityParameters {
-                Target = GameManager.Instance.CommandManager.EntitiesOnGrid.GetEntityByID((int)json["Target"])
+                Target = GameManager.Instance.CommandManager.EntitiesOnGrid.GetEntityByID((long)json["Target"])
             };
         }
     }
